+++ assembly/src/ast/instruction/advice.rs
@@ -12,10 +12,6 @@ use vm_core::sys_events::SystemEvent;
 /// - Insert new data into the advice map.
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub enum SystemEventNode {
-    PushU64Div,
-    PushFalconDiv,
-    PushExt2intt,
-    PushSmtPeek,
     PushMapVal,
     PushMapValN,
     PushMtNode,
@@ -29,10 +25,6 @@ impl From<&SystemEventNode> for SystemEvent {
     fn from(value: &SystemEventNode) -> Self {
         use SystemEventNode::*;
         match value {
-            PushU64Div => Self::U64Div,
-            PushFalconDiv => Self::FalconDiv,
-            PushExt2intt => Self::Ext2Intt,
-            PushSmtPeek => Self::SmtPeek,
             PushMapVal => Self::MapValueToStack,
             PushMapValN => Self::MapValueToStackN,
             PushMtNode => Self::MerkleNodeToStack,
@@ -53,10 +45,6 @@ impl crate::prettier::PrettyPrint for SystemEventNode {
 impl fmt::Display for SystemEventNode {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::PushU64Div => write!(f, "push_u64div"),
-            Self::PushFalconDiv => write!(f, "push_falcon_div"),
-            Self::PushExt2intt => write!(f, "push_ext2intt"),
-            Self::PushSmtPeek => write!(f, "push_smtpeek"),
             Self::PushMapVal => write!(f, "push_mapval"),
             Self::PushMapValN => write!(f, "push_mapvaln"),
             Self::PushMtNode => write!(f, "push_mtnode"),