+++ processor/src/operations/sys_ops/sys_event_handlers.rs
@@ -1,25 +1,15 @@
 use alloc::vec::Vec;
 
 use vm_core::{
-    crypto::{
-        hash::{Rpo256, RpoDigest},
-        merkle::{EmptySubtreeRoots, Smt, SMT_DEPTH},
-    },
+    crypto::hash::{Rpo256, RpoDigest},
     sys_events::SystemEvent,
-    Felt, FieldElement, Word, WORD_SIZE, ZERO,
+    Felt, FieldElement, WORD_SIZE, ZERO,
 };
-use winter_prover::math::fft;
 
-use crate::{
-    AdviceProvider, AdviceSource, ExecutionError, Ext2InttError, Host, Process, ProcessState,
-    QuadFelt,
-};
+use crate::{AdviceProvider, AdviceSource, ExecutionError, Host, Process, ProcessState, QuadFelt};
 
 /// The offset of the domain value on the stack in the `hdword_to_map_with_domain` system event.
 const HDWORD_TO_MAP_WITH_DOMAIN_DOMAIN_OFFSET: usize = 8;
-
-/// Falcon signature prime.
-const M: u64 = 12289;
 
 impl Process {
     pub(super) fn handle_system_event(
@@ -40,11 +30,7 @@ impl Process {
             SystemEvent::MapValueToStackN => {
                 copy_map_value_to_adv_stack(advice_provider, process_state, true)
             },
-            SystemEvent::U64Div => push_u64_div_result(advice_provider, process_state),
-            SystemEvent::FalconDiv => push_falcon_mod_result(advice_provider, process_state),
             SystemEvent::Ext2Inv => push_ext2_inv_result(advice_provider, process_state),
-            SystemEvent::Ext2Intt => push_ext2_intt_result(advice_provider, process_state),
-            SystemEvent::SmtPeek => push_smtpeek_result(advice_provider, process_state),
             SystemEvent::U32Clz => push_leading_zeros(advice_provider, process_state),
             SystemEvent::U32Ctz => push_trailing_zeros(advice_provider, process_state),
             SystemEvent::U32Clo => push_leading_ones(advice_provider, process_state),
@@ -294,92 +280,6 @@ pub fn copy_map_value_to_adv_stack(
     Ok(())
 }
 
-/// Pushes the result of [u64] division (both the quotient and the remainder) onto the advice
-/// stack.
-///
-/// Inputs:
-///   Operand stack: [b1, b0, a1, a0, ...]
-///   Advice stack: [...]
-///
-/// Outputs:
-///   Operand stack: [b1, b0, a1, a0, ...]
-///   Advice stack: [q0, q1, r0, r1, ...]
-///
-/// Where (a0, a1) and (b0, b1) are the 32-bit limbs of the dividend and the divisor
-/// respectively (with a0 representing the 32 lest significant bits and a1 representing the
-/// 32 most significant bits). Similarly, (q0, q1) and (r0, r1) represent the quotient and
-/// the remainder respectively.
-///
-/// # Errors
-/// Returns an error if the divisor is ZERO.
-pub fn push_u64_div_result(
-    advice_provider: &mut impl AdviceProvider,
-    process: ProcessState,
-) -> Result<(), ExecutionError> {
-    let divisor_hi = process.get_stack_item(0).as_int();
-    let divisor_lo = process.get_stack_item(1).as_int();
-    let divisor = (divisor_hi << 32) + divisor_lo;
-
-    if divisor == 0 {
-        return Err(ExecutionError::DivideByZero(process.clk()));
-    }
-
-    let dividend_hi = process.get_stack_item(2).as_int();
-    let dividend_lo = process.get_stack_item(3).as_int();
-    let dividend = (dividend_hi << 32) + dividend_lo;
-
-    let quotient = dividend / divisor;
-    let remainder = dividend - quotient * divisor;
-
-    let (q_hi, q_lo) = u64_to_u32_elements(quotient);
-    let (r_hi, r_lo) = u64_to_u32_elements(remainder);
-
-    advice_provider.push_stack(AdviceSource::Value(r_hi))?;
-    advice_provider.push_stack(AdviceSource::Value(r_lo))?;
-    advice_provider.push_stack(AdviceSource::Value(q_hi))?;
-    advice_provider.push_stack(AdviceSource::Value(q_lo))?;
-
-    Ok(())
-}
-
-/// Pushes the result of divison (both the quotient and the remainder) of a [u64] by the Falcon
-/// prime (M = 12289) onto the advice stack.
-///
-/// Inputs:
-///   Operand stack: [a1, a0, ...]
-///   Advice stack: [...]
-///
-/// Outputs:
-///   Operand stack: [a1, a0, ...]
-///   Advice stack: [q1, q0, r, ...]
-///
-/// Where (a0, a1) are the 32-bit limbs of the dividend (with a0 representing the 32 least
-/// significant bits and a1 representing the 32 most significant bits).
-/// Similarly, (q0, q1) represent the quotient and r the remainder.
-///
-/// # Errors
-/// Returns an error if the divisor is ZERO.
-pub fn push_falcon_mod_result(
-    advice_provider: &mut impl AdviceProvider,
-    process: ProcessState,
-) -> Result<(), ExecutionError> {
-    let dividend_hi = process.get_stack_item(0).as_int();
-    let dividend_lo = process.get_stack_item(1).as_int();
-    let dividend = (dividend_hi << 32) + dividend_lo;
-
-    let (quotient, remainder) = (dividend / M, dividend % M);
-
-    let (q_hi, q_lo) = u64_to_u32_elements(quotient);
-    let (r_hi, r_lo) = u64_to_u32_elements(remainder);
-    assert_eq!(r_hi, ZERO);
-
-    advice_provider.push_stack(AdviceSource::Value(r_lo))?;
-    advice_provider.push_stack(AdviceSource::Value(q_lo))?;
-    advice_provider.push_stack(AdviceSource::Value(q_hi))?;
-
-    Ok(())
-}
-
 /// Given an element in a quadratic extension field on the top of the stack (i.e., a0, b1),
 /// computes its multiplicative inverse and push the result onto the advice stack.
 ///
@@ -411,90 +311,6 @@ pub fn push_ext2_inv_result(
 
     advice_provider.push_stack(AdviceSource::Value(result[1]))?;
     advice_provider.push_stack(AdviceSource::Value(result[0]))?;
-
-    Ok(())
-}
-
-/// Given evaluations of a polynomial over some specified domain, interpolates the evaluations
-///  into a polynomial in coefficient form and pushes the result into the advice stack.
-///
-/// The interpolation is performed using the iNTT algorithm. The evaluations are expected to be
-/// in the quadratic extension.
-///
-/// Inputs:
-///   Operand stack: [output_size, input_size, input_start_ptr, ...]
-///   Advice stack: [...]
-///
-/// Outputs:
-///   Operand stack: [output_size, input_size, input_start_ptr, ...]
-///   Advice stack: [coefficients...]
-///
-/// - `input_size` is the number of evaluations (each evaluation is 2 base field elements). Must be
-///   a power of 2 and greater 1.
-/// - `output_size` is the number of coefficients in the interpolated polynomial (each coefficient
-///   is 2 base field elements). Must be smaller than or equal to the number of input evaluations.
-/// - `input_start_ptr` is the memory address of the first evaluation.
-/// - `coefficients` are the coefficients of the interpolated polynomial such that lowest degree
-///   coefficients are located at the top of the advice stack.
-///
-/// # Errors
-/// Returns an error if:
-/// - `input_size` less than or equal to 1, or is not a power of 2.
-/// - `output_size` is 0 or is greater than the `input_size`.
-/// - `input_ptr` is greater than 2^32, or is not aligned on a word boundary.
-/// - `input_ptr + input_size * 2` is greater than 2^32.
-pub fn push_ext2_intt_result(
-    advice_provider: &mut impl AdviceProvider,
-    process: ProcessState,
-) -> Result<(), ExecutionError> {
-    let output_size = process.get_stack_item(0).as_int() as usize;
-    let input_size = process.get_stack_item(1).as_int() as usize;
-    let input_start_ptr = process.get_stack_item(2).as_int();
-
-    if input_size <= 1 {
-        return Err(Ext2InttError::DomainSizeTooSmall(input_size as u64).into());
-    }
-    if !input_size.is_power_of_two() {
-        return Err(Ext2InttError::DomainSizeNotPowerOf2(input_size as u64).into());
-    }
-    if input_start_ptr >= u32::MAX as u64 {
-        return Err(Ext2InttError::InputStartAddressTooBig(input_start_ptr).into());
-    }
-    if input_start_ptr % WORD_SIZE as u64 != 0 {
-        return Err(Ext2InttError::InputStartNotWordAligned(input_start_ptr).into());
-    }
-    if input_size > u32::MAX as usize {
-        return Err(Ext2InttError::InputSizeTooBig(input_size as u64).into());
-    }
-
-    let input_end_ptr = input_start_ptr + (input_size * 2) as u64;
-    if input_end_ptr > u32::MAX as u64 {
-        return Err(Ext2InttError::InputEndAddressTooBig(input_end_ptr).into());
-    }
-
-    if output_size == 0 {
-        return Err(Ext2InttError::OutputSizeIsZero.into());
-    }
-    if output_size > input_size {
-        return Err(Ext2InttError::OutputSizeTooBig(output_size, input_size).into());
-    }
-
-    let mut poly = Vec::with_capacity(input_size);
-    for addr in ((input_start_ptr as u32)..(input_end_ptr as u32)).step_by(4) {
-        let word = process
-            .get_mem_word(process.ctx(), addr)?
-            .ok_or(Ext2InttError::UninitializedMemoryAddress(addr))?;
-
-        poly.push(QuadFelt::new(word[0], word[1]));
-        poly.push(QuadFelt::new(word[2], word[3]));
-    }
-
-    let twiddles = fft::get_inv_twiddles::<Felt>(input_size);
-    fft::interpolate_poly::<Felt, QuadFelt>(&mut poly, &twiddles);
-
-    for element in QuadFelt::slice_as_base_elements(&poly[..output_size]).iter().rev() {
-        advice_provider.push_stack(AdviceSource::Value(*element))?;
-    }
 
     Ok(())
 }
@@ -595,61 +411,6 @@ pub fn push_ilog2(
     Ok(())
 }
 
-/// Pushes onto the advice stack the value associated with the specified key in a Sparse
-/// Merkle Tree defined by the specified root.
-///
-/// If no value was previously associated with the specified key, [ZERO; 4] is pushed onto
-/// the advice stack.
-///
-/// Inputs:
-///   Operand stack: [KEY, ROOT, ...]
-///   Advice stack: [...]
-///
-/// Outputs:
-///   Operand stack: [KEY, ROOT, ...]
-///   Advice stack: [VALUE, ...]
-///
-/// # Errors
-/// Returns an error if the provided Merkle root doesn't exist on the advice provider.
-///
-/// # Panics
-/// Will panic as unimplemented if the target depth is `64`.
-pub fn push_smtpeek_result(
-    advice_provider: &mut impl AdviceProvider,
-    process: ProcessState,
-) -> Result<(), ExecutionError> {
-    let empty_leaf = EmptySubtreeRoots::entry(SMT_DEPTH, SMT_DEPTH);
-    // fetch the arguments from the operand stack
-    let key = process.get_stack_word(0);
-    let root = process.get_stack_word(1);
-
-    // get the node from the SMT for the specified key; this node can be either a leaf node,
-    // or a root of an empty subtree at the returned depth
-    let node = advice_provider.get_tree_node(root, &Felt::new(SMT_DEPTH as u64), &key[3])?;
-
-    if node == Word::from(empty_leaf) {
-        // if the node is a root of an empty subtree, then there is no value associated with
-        // the specified key
-        advice_provider.push_stack(AdviceSource::Word(Smt::EMPTY_VALUE))?;
-    } else {
-        let leaf_preimage = get_smt_leaf_preimage(advice_provider, node)?;
-
-        for (key_in_leaf, value_in_leaf) in leaf_preimage {
-            if key == key_in_leaf {
-                // Found key - push value associated with key, and return
-                advice_provider.push_stack(AdviceSource::Word(value_in_leaf))?;
-
-                return Ok(());
-            }
-        }
-
-        // if we can't find any key in the leaf that matches `key`, it means no value is
-        // associated with `key`
-        advice_provider.push_stack(AdviceSource::Word(Smt::EMPTY_VALUE))?;
-    }
-    Ok(())
-}
-
 // HELPER METHODS
 // --------------------------------------------------------------------------------------------
 
@@ -677,12 +438,6 @@ fn get_mem_addr_range(
     Ok((start_addr as u32, end_addr as u32))
 }
 
-fn u64_to_u32_elements(value: u64) -> (Felt, Felt) {
-    let hi = Felt::from((value >> 32) as u32);
-    let lo = Felt::from(value as u32);
-    (hi, lo)
-}
-
 /// Gets the top stack element, applies a provided function to it and pushes it to the advice
 /// provider.
 fn push_transformed_stack_top<A: AdviceProvider>(
@@ -698,29 +453,4 @@ fn push_transformed_stack_top<A: AdviceProvider>(
     let transformed_stack_top = f(stack_top);
     advice_provider.push_stack(AdviceSource::Value(transformed_stack_top))?;
     Ok(())
-}
-
-fn get_smt_leaf_preimage<A: AdviceProvider>(
-    advice_provider: &A,
-    node: Word,
-) -> Result<Vec<(Word, Word)>, ExecutionError> {
-    let node_bytes = RpoDigest::from(node);
-
-    let kv_pairs = advice_provider
-        .get_mapped_values(&node_bytes)
-        .ok_or(ExecutionError::SmtNodeNotFound(node))?;
-
-    if kv_pairs.len() % WORD_SIZE * 2 != 0 {
-        return Err(ExecutionError::SmtNodePreImageNotValid(node, kv_pairs.len()));
-    }
-
-    Ok(kv_pairs
-        .chunks_exact(WORD_SIZE * 2)
-        .map(|kv_chunk| {
-            let key = [kv_chunk[0], kv_chunk[1], kv_chunk[2], kv_chunk[3]];
-            let value = [kv_chunk[4], kv_chunk[5], kv_chunk[6], kv_chunk[7]];
-
-            (key, value)
-        })
-        .collect())
 }